salesforce/dataloader/airflow/requirements.txt
azure-storage-blob
salesforce-bulk
pandas
faker
petname
